tstime: add an abstract Time interface, which can be real or virtual.
The virtual implementation supports configurable "time dilation", to
make it run faster or slower than virtual time.
